pyproject.toml
[project]
name = "vp-bot"
version = "0.1.0"
description = "Бот для города Вятские Поляны"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot[job-queue]>=22.5",
    "sqlmodel>=0.0.31",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
